FinalAssessment/requirements.txt
torch
torchvision
numpy
opencv-python
albumentations
onnx
onnxruntime-gpu
onnxconverter-common
gradio
tqdm